Some Modified

+++ resources/views/daily_workouts/table.blade.php
@@ -1,3 +1,4 @@
+
 <div class="table-responsive">
     <table class="table" id="dailyWorkouts-table">
         <thead>
@@ -27,14 +28,16 @@
             <td>{{ $dailyWorkouts->duration_minutes }}</td>
             <td>{{ $dailyWorkouts->created_at }}</td>
                 <td>
-                    {!! Form::open(['route' => ['dailyWorkouts.destroy', $dailyWorkouts->id], 'method' => 'delete']) !!}
-                    <div class='btn-group'>
-                        <a href="{{ route('dailyWorkouts.show', [$dailyWorkouts->id]) }}" class='btn btn-outline-primary btn-xs'><i class="im im-icon-Eye" data-placement="top" title="View"></i></a>
-                        <a href="{{ route('dailyWorkouts.edit', [$dailyWorkouts->id]) }}" class='btn btn-outline-primary btn-xs'><i
-                                class="im im-icon-Pen"  data-toggle="tooltip" data-placement="top" title="Edit"></i></a>
-                        {!! Form::button('<i class="im im-icon-Remove" data-toggle="tooltip" data-placement="top" title="Delete"></i>', ['type' => 'submit', 'class' => 'btn btn-outline-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
-                    </div>
-                    {!! Form::close() !!}
+                    <a href="{{ route('dailyWorkouts.show', [$dailyWorkouts->id]) }}" class='btn btn-outline-primary btn-xs'><i class="im im-icon-Eye" data-placement="top" title="View"></i></a>
+                    @if(if_can('show_all_data'))
+                        {!! Form::open(['route' => ['dailyWorkouts.destroy', $dailyWorkouts->id], 'method' => 'delete']) !!}
+                        <div class='btn-group'>
+                            <a href="{{ route('dailyWorkouts.edit', [$dailyWorkouts->id]) }}" class='btn btn-outline-primary btn-xs'><i
+                                    class="im im-icon-Pen"  data-toggle="tooltip" data-placement="top" title="Edit"></i></a>
+                            {!! Form::button('<i class="im im-icon-Remove" data-toggle="tooltip" data-placement="top" title="Delete"></i>', ['type' => 'submit', 'class' => 'btn btn-outline-danger btn-xs', 'onclick' => "return confirm('Are you sure?')"]) !!}
+                        </div>
+                        {!! Form::close() !!}
+                    @endif
                 </td>
             </tr>
         @endforeach
